requirements.txt
nextcord[voice]==2.6.0
python-dotenv==1.0.0
langchain==0.1.11
langsmith
langchain_openai
tiktoken==0.5.2
langchain-google-genai
transformers==4.36.2
pydantic_settings==2.1.0
pillow==10.1.0
wikipedia==1.4.0
pyowm==3.3.0
google-api-python-client==2.111.0
python-telegram-bot[ext]==20.7
langchain_mistralai
setuptools
langchain-groq
langchain-anthropic
types-beautifulsoup4
